samples/python/agents/images/pyproject.toml
[project]
name = "images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.2",
    "crewai[tools]>=0.165.1",
    "google-genai>=1.31.0",
    "langsmith>=0.3.45",
    "litellm>=1.0.0",
    "python-dotenv>=1.1.1",
    "starlette>=0.47.2",
]


[project.scripts]
images = "app.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
